+++ resource_manager.hpp
@@ -6,34 +6,35 @@ namespace tano
 {
 #if WITH_UNPACKED_RESOUCES
 
-  typedef function<bool (const string&, void *)> cbFileChanged;
+  typedef function<bool (const string&, void* )> cbFileChanged;
 
   class ResourceManager
   {
   public:
-    ResourceManager(const char *outputFilename);
+    ResourceManager(const char* outputFilename);
     ~ResourceManager();
 
     static ResourceManager &Instance();
-    static bool Create(const char *outputFilename);
+    static bool Create(const char* outputFilename);
     static bool Destroy();
 
-    bool FileExists(const char *filename);
-    __time64_t ModifiedDate(const char *filename);
-    bool LoadFile(const char *filename, vector<char> *buf);
-    bool LoadPartial(const char *filename, size_t ofs, size_t len, vector<char> *buf);
-    bool LoadInplace(const char *filename, size_t ofs, size_t len, void *buf);
+    bool FileExists(const char* filename);
+    __time64_t ModifiedDate(const char* filename);
+    bool LoadFile(const char* filename, vector<char>* buf);
+    bool LoadPartial(const char* filename, u32 ofs, u32 len, vector<char>* buf);
+    bool LoadInplace(const char* filename, u32 ofs, u32 len, void* buf);
+
     ObjectHandle LoadTexture(
         const char* filename,
         const char* friendlyName = nullptr,
         bool srgb = false,
-        D3DX11_IMAGE_INFO *info = nullptr);
+        D3DX11_IMAGE_INFO* info = nullptr);
     ObjectHandle LoadTextureFromMemory(
         const char* buf,
-        size_t len,
+        u32 len,
         const char* friendlyName,
         bool srgb,
-        D3DX11_IMAGE_INFO *info);
+        D3DX11_IMAGE_INFO* info);
 
     void AddPath(const string &path);
 
@@ -49,7 +50,7 @@ namespace tano
     void Tick();
 
   private:
-    string ResolveFilename(const char *filename, bool fullPath);
+    string ResolveFilename(const char* filename, bool fullPath);
 
     FileWatcher _fileWatcher;
 
@@ -81,24 +82,24 @@ namespace tano
   class PackedResourceManager
   {
   public:
-    PackedResourceManager(const char *resourceFile);
+    PackedResourceManager(const char* resourceFile);
     ~PackedResourceManager();
 
     static PackedResourceManager &instance();
-    static bool create(const char *resourceFile);
+    static bool create(const char* resourceFile);
     static bool close();
 
-    bool load_file(const char *filename, vector<char> *buf);
-    bool load_partial(const char *filename, size_t ofs, size_t len, vector<char> *buf);
-    bool load_inplace(const char *filename, size_t ofs, size_t len, void *buf);
-    ObjectHandle LoadTexture(const char *filename, const char *friendly_name, bool srgb, D3DX11_IMAGE_INFO *info);
-    ObjectHandle LoadTextureFromMemory(const char *buf, size_t len, const char *friendly_name, bool srgb, D3DX11_IMAGE_INFO *info);
+    bool load_file(const char* filename, vector<char>* buf);
+    bool load_partial(const char* filename, size_t ofs, size_t len, vector<char>* buf);
+    bool load_inplace(const char* filename, size_t ofs, size_t len, void* buf);
+    ObjectHandle LoadTexture(const char* filename, const char* friendly_name, bool srgb, D3DX11_IMAGE_INFO* info);
+    ObjectHandle LoadTextureFromMemory(const char* buf, size_t len, const char* friendly_name, bool srgb, D3DX11_IMAGE_INFO* info);
 
   private:
 
-    bool loadPackedFile(const char *filename, vector<char> *buf);
-    bool loadPackedInplace(const char *filename, size_t ofs, size_t len, void *buf);
-    int hashLookup(const char *key);
+    bool loadPackedFile(const char* filename, vector<char>* buf);
+    bool loadPackedInplace(const char* filename, size_t ofs, size_t len, void* buf);
+    int hashLookup(const char* key);
 
     struct PackedFileInfo {
       int offset;
@@ -111,7 +112,7 @@ namespace tano
     vector<int> _finalHash;
     vector<PackedFileInfo> _fileInfo;
 
-    static PackedResourceManager *_instance;
+    static PackedResourceManager* _instance;
     string _resourceFile;
   };
